Fix failed test & moved validation into model
