+++ app/api/permissions/entitiesPermissions.ts
@@ -4,7 +4,12 @@ import userGroups from 'api/usergroups/userGroups';
 import { unique } from 'api/utils/filters';
 import { GroupMemberSchema, UserGroupSchema } from 'shared/types/userGroupType';
 import { EntitySchema } from 'shared/types/entityType';
-import { AccessLevels, PermissionType, MixedAccess } from 'shared/types/permissionSchema';
+import {
+  AccessLevels,
+  PermissionType,
+  MixedAccess,
+  validateUniquePermissions,
+} from 'shared/types/permissionSchema';
 import { PermissionSchema } from 'shared/types/permissionType';
 import { MemberWithPermission } from 'shared/types/entityPermisions';
 
@@ -49,7 +54,9 @@ async function setAccessLevelAndPermissionData(
 
 export const entitiesPermissions = {
   set: async (permissionsData: any) => {
+    await validateUniquePermissions(permissionsData);
     const currentEntities = await entities.get({ sharedId: { $in: permissionsData.ids } }, '_id');
+
     const toSave = currentEntities.map(entity => ({
       _id: entity._id,
       permissions: permissionsData.permissions,
